+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added `hoursOnly`, `minutesOnly`, `secondsOnly`, `millisecondsOnly`, and
   `microsecondsOnly` extension getters to `Duration`.
+* Added a `tryParseDuration` function.
 
 ## 0.5.0